+++ src/main/java/com/sparta/springscheduleappnew/service/ScheduleService.java
@@ -4,6 +4,8 @@ import com.sparta.springscheduleappnew.dto.ScheduleRequestDto;
 import com.sparta.springscheduleappnew.dto.ScheduleResponseDto;
 import com.sparta.springscheduleappnew.entity.Schedule;
 import com.sparta.springscheduleappnew.entity.User;
+import com.sparta.springscheduleappnew.errors.CustomException;
+import com.sparta.springscheduleappnew.errors.ErrorCode;
 import com.sparta.springscheduleappnew.repository.ScheduleRepository;
 import com.sparta.springscheduleappnew.repository.UserRepository;
 import java.util.Optional;
@@ -21,9 +23,9 @@ public class ScheduleService {
     private final UserRepository userRepository;
 
     public ScheduleResponseDto createSchedule(ScheduleRequestDto scheduleDto, Long authorId) {
-        // authorId로 사용자(작성자)를 검색
-        User author = userRepository.findById(authorId)
-                .orElseThrow(() -> new RuntimeException("작성자가 존재하지 않습니다."));
+    // authorId로 사용자(작성자)를 검색
+    User author = userRepository.findById(authorId)
+            .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
 
         // 일정의 생성과 저장
         Schedule schedule = Schedule.builder()
@@ -36,9 +38,11 @@ public class ScheduleService {
         return new ScheduleResponseDto(savedSchedule);
     }
 
-    public Optional<ScheduleResponseDto> getScheduleById(Long id) {
-        return scheduleRepository.findById(id)
-                .map(ScheduleResponseDto::new);
+    public ScheduleResponseDto getScheduleById(Long id) {
+        Schedule schedule = scheduleRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ErrorCode.SCHEDULE_NOT_FOUND));
+
+        return new ScheduleResponseDto(schedule);
     }
 
     public Page<ScheduleResponseDto> getSchedules(int page, int size) {
@@ -48,8 +52,8 @@ public class ScheduleService {
     }
 
     public ScheduleResponseDto updateSchedule(Long id, ScheduleRequestDto scheduleDto) {
-        Schedule schedule = scheduleRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("일정이 존재하지 않습니다."));
+    Schedule schedule = scheduleRepository.findById(id)
+            .orElseThrow(() -> new CustomException(ErrorCode.SCHEDULE_NOT_FOUND));
 
         schedule.setTitle(scheduleDto.getTitle());
         schedule.setDescription(scheduleDto.getDescription());
@@ -59,6 +63,9 @@ public class ScheduleService {
     }
 
     public void deleteSchedule(Long id) {
+        if (!scheduleRepository.existsById(id)) {
+            throw new CustomException(ErrorCode.SCHEDULE_NOT_FOUND);
+        }
         scheduleRepository.deleteById(id);
     }
 }
